Introduce reusable closable notebook
